refactor(docs): scope type:module to lib/ and extract llm-views
Addresses peqy review note about "type": "module" being a
package-wide change. The previous commit set it at docs/src/
level, which under strict Node ESM rules implicitly disallows
extensionless relative imports — and there are ~40 such imports
across app/ and components/ today. Next.js's bundler doesn't
enforce ESM extension rules so the build itself works fine, but
the package-wide flag is a latent foot-gun for any future tool
that loads these files via plain Node ESM.

This refactor moves "type": "module" to docs/src/lib/package.json
so only files under lib/ are subject to ESM strict mode. To make
the route handlers' logic testable without needing app/ to be
ESM, the rendering for /llms.txt, /llms-full.txt, and /<page>.md
is extracted into lib/llm-views.js:

  - renderLlmsTxt()        builds the curated index body
  - renderLlmsFullTxt()    builds the full corpus body
  - renderPageMarkdown()   returns { status, body } for one page

Route handlers in app/ become 10-30 line Response wrappers around
these helpers. Tests now import the lib functions directly, no
longer pulling Next.js-bundled route files.

Verified:
- npm test: 41/41 pass
- npx eslint . --max-warnings 0: clean
- npx prettier --check .: clean
- npx next build: compiles successfully (the unrelated _meta
  "explorer" validation failure is pre-existing on main and
  unrelated to this PR).
