+++ docs/src/lib/__tests__/llms-txt.test.js
@@ -1,5 +1,9 @@
 /**
- * Integration smoke test for the /llms.txt route.
+ * Smoke tests for the renderLlmsTxt() / renderLlmsFullTxt() lib helpers.
+ *
+ * These functions back the /llms.txt and /llms-full.txt routes; the route
+ * handlers themselves are thin Response wrappers so we exercise the logic
+ * here at the lib level (no Next.js runtime required).
  *
  * Asserts the contract that LLM-ingestion clients depend on:
  *   - Header: H1 site name + blockquote summary
@@ -7,60 +11,49 @@
  *   - Every page link points to the .md variant
  *   - "## Optional" block is present when glossary/changelog exist
  *   - Resources footer is present
+ *   - llms-full.txt has no double "---" page boundary
  *
- * This test imports the route handler module directly and invokes its GET
- * function — it does NOT spin up a Next.js dev server. The route depends on
- * the real content directory, so this is a true end-to-end smoke test.
- *
- * If the docs/content tree is unavailable (e.g. in a stripped CI image),
- * this test soft-skips rather than failing.
+ * Soft-skips when docs/content is unavailable.
  */
 
 import { test } from 'node:test'
 import assert from 'node:assert/strict'
 
 import { findContentDir, clearPagesCache } from '../content-index.js'
-import { GET as getLlmsTxt } from '../../app/llms.txt/route.js'
+import { renderLlmsTxt, renderLlmsFullTxt } from '../llm-views.js'
 
 const hasContent = !!findContentDir()
 
-test('GET /llms.txt: responds with text/plain and a non-empty body', async t => {
+test('renderLlmsTxt: returns a non-empty string', t => {
   if (!hasContent) {
     t.skip('docs/content directory not present; skipping integration smoke test')
     return
   }
   clearPagesCache()
-  const res = await getLlmsTxt()
-  assert.equal(res.status, 200)
-  assert.match(res.headers.get('Content-Type') || '', /^text\/plain/)
-  const body = await res.text()
+  const body = renderLlmsTxt()
+  assert.equal(typeof body, 'string')
   assert.ok(body.length > 200, 'body should contain a real index')
 })
 
-test('GET /llms.txt: includes header and blockquote summary', async t => {
+test('renderLlmsTxt: includes header and blockquote summary', t => {
   if (!hasContent) {
     t.skip('docs/content directory not present')
     return
   }
   clearPagesCache()
-  const res = await getLlmsTxt()
-  const body = await res.text()
+  const body = renderLlmsTxt()
   assert.match(body, /^# Rhesis Documentation\b/m)
   assert.match(body, /^>\s+\S+/m)
 })
 
-test('GET /llms.txt: every link points to a .md URL', async t => {
+test('renderLlmsTxt: every internal link points to a .md URL', t => {
   if (!hasContent) {
     t.skip('docs/content directory not present')
     return
   }
   clearPagesCache()
-  const res = await getLlmsTxt()
-  const body = await res.text()
+  const body = renderLlmsTxt()
 
-  // Find every markdown link that targets a docs.rhesis.ai URL. Filter out
-  // external links (GitHub, Discord, the readthedocs SDK reference, the
-  // llms-full.txt resource link).
   const linkRegex = /\[[^\]]+\]\((https:\/\/docs\.rhesis\.ai\/[^)]+)\)/g
   const urls = []
   for (const m of body.matchAll(linkRegex)) {
@@ -74,19 +67,17 @@ test('GET /llms.txt: every link points to a .md URL', async t => {
   }
 })
 
-test('GET /llms.txt: includes the expected primary section headings', async t => {
+test('renderLlmsTxt: includes the expected primary section headings', t => {
   if (!hasContent) {
     t.skip('docs/content directory not present')
     return
   }
   clearPagesCache()
-  const res = await getLlmsTxt()
-  const body = await res.text()
+  const body = renderLlmsTxt()
 
-  // Sections are conditional on having pages; at minimum we expect Docs.
+  // At minimum we expect Docs.
   assert.match(body, /^## Docs$/m, 'Docs section should be present')
 
-  // If Optional exists, it should appear after the primary sections.
   if (/^## Optional$/m.test(body)) {
     const optionalIdx = body.search(/^## Optional$/m)
     const docsIdx = body.search(/^## Docs$/m)
@@ -94,14 +85,48 @@ test('GET /llms.txt: includes the expected primary section headings', async t =>
   }
 })
 
-test('GET /llms.txt: includes Resources footer with llms-full.txt pointer', async t => {
+test('renderLlmsTxt: includes Resources footer with llms-full.txt pointer', t => {
   if (!hasContent) {
     t.skip('docs/content directory not present')
     return
   }
   clearPagesCache()
-  const res = await getLlmsTxt()
-  const body = await res.text()
+  const body = renderLlmsTxt()
   assert.match(body, /^## Resources$/m)
   assert.match(body, /llms-full\.txt/)
+})
+
+// ---------------------------------------------------------------------------
+// renderLlmsFullTxt
+// ---------------------------------------------------------------------------
+
+test('renderLlmsFullTxt: returns a substantial markdown corpus', t => {
+  if (!hasContent) {
+    t.skip('docs/content directory not present')
+    return
+  }
+  clearPagesCache()
+  const body = renderLlmsFullTxt()
+  assert.equal(typeof body, 'string')
+  // The corpus is ~1 MB in production. Use a low floor so the test is robust
+  // to content tree size changes.
+  assert.ok(body.length > 5000, `corpus should be substantial, got ${body.length} bytes`)
+  assert.match(body, /^# Rhesis Documentation \(full\)$/m)
+})
+
+test('renderLlmsFullTxt: pages are delimited by frontmatter (no double "---")', t => {
+  if (!hasContent) {
+    t.skip('docs/content directory not present')
+    return
+  }
+  clearPagesCache()
+  const body = renderLlmsFullTxt()
+  // An empty fence (---\n\s*\n---) with no key/value rows would mean we still
+  // have the redundant separator. Frontmatter blocks always start with
+  // "url:" or "title:" on the next line, so this regex catches the bug.
+  const emptyFence = /\n---\n\s*\n---\n(?!url:|title:)/
+  assert.ok(
+    !emptyFence.test(body),
+    'llms-full.txt should not contain empty/redundant --- separators between pages'
+  )
 })
